Move `sha256Bytes` util functions to cipher suite
