+++ libs/layered-property-map/src/test/kotlin/net/corda/layeredpropertymap/LayeredPropertyMapTest.kt
@@ -1,6 +1,7 @@
 package net.corda.layeredpropertymap
 
 import net.corda.crypto.cipher.suite.PublicKeyHash
+import net.corda.crypto.cipher.suite.sha256Bytes
 import net.corda.layeredpropertymap.impl.LayeredPropertyMapImpl
 import net.corda.layeredpropertymap.impl.PropertyConverter
 import net.corda.test.util.createTestCase
@@ -11,7 +12,6 @@ import net.corda.utilities.parseSet
 import net.corda.v5.base.exceptions.ValueNotFoundException
 import net.corda.v5.base.types.ByteArrays.toHexString
 import net.corda.v5.base.types.LayeredPropertyMap
-import net.corda.v5.crypto.sha256Bytes
 import org.junit.jupiter.api.Test
 import java.time.Instant
 import java.util.concurrent.ConcurrentHashMap
